requirements.txt
APScheduler==3.11.2
python-dotenv==1.2.1
slack_bolt==1.27.0
slack_sdk==3.40.1
tzlocal==5.3.1
